pyproject.toml
[project]
name = "lld-proj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "brotli>=1.1.0",
    "bs4>=0.0.2",
    "click>=8.2.1",
    "faiss-cpu>=1.11.0",
    "fake-useragent>=2.2.0",
    "fastapi>=0.115.12",
    "langchain-groq>=0.3.2",
    "langchain>=0.3.25",
    "linkedin-jobs-scraper>=5.0.2",
    "loguru>=0.7.3",
    "numpy>=2.3.0",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "torch>=2.7.1",
    "uvicorn>=0.34.3",
    "webdriver-manager>=4.0.2",
    "rich>=14.0.0",
]
